move update-specs.js
